==> 01.0_metab_reformat_param_file.R <==
## Parameter file for Metabolon metabolite reformating
## by: Laura Corbin

project=direct1
dataset=metabolon
processeddatadirectory=U:\Data\Processed\
scriptsdirectory=U:\Scripts\
fullpath_2_excel_file=P:\
path_to_metabolon=X:\Metabolon\
externaldatadirectory=U:\Data\ExternalData\
