==> 01.0_nmr_reformat_param_file.R <==
## Paramater file for NMR metabolite reformating
## by: Laura Corbin

project=direct1
dataset=nmr
processeddatadirectory=U:\Data\Processed\
scriptsdirectory=U:\Scripts\
fullpath_2_excel_file=X:\NMR\
n_feature_meta_rows=12
n_sample_meta_col=5
feature_id_row=10
sample_id_col=E
last_sample=586
last_feature=HX
filename=DIGL - Results.xlsx
