==> 03.0_pheno_data_prep_param_file.R <==
## Parameter file for phenodata prep
## by: Laura Corbin

project=direct1
processeddatadirectory=U:\Data\Processed\
scriptsdirectory=U:\Scripts\
rawdata_dir=P:\
id_correction="REDACTED4,REDACTED5"
exclusions="REDACTED6,REDACTED7"
